+++ rollup.config.js
@@ -5,13 +5,29 @@ import svelte from 'rollup-plugin-svelte';
 import babel from '@rollup/plugin-babel';
 import { terser } from 'rollup-plugin-terser';
 import config from 'sapper/config/rollup.js';
+import json from '@rollup/plugin-json';
+
 import pkg from './package.json';
 
 const mode = process.env.NODE_ENV;
 const dev = mode === 'development';
 const legacy = !!process.env.SAPPER_LEGACY_BUILD;
 
-const onwarn = (warning, onwarn) => (warning.code === 'CIRCULAR_DEPENDENCY' && /[/\\]@sapper[/\\]/.test(warning.message)) || onwarn(warning);
+const onwarn = (warning, onwarn) => {
+	if (
+	  (warning.code === 'CIRCULAR_DEPENDENCY' &&
+		/[/\\]@sapper[/\\]/.test(warning.message))
+	) {
+	  return
+	}
+  
+	// ignores the annoying this is undefined warning
+	if(warning.code === 'THIS_IS_UNDEFINED') {
+	  return
+	}
+  
+	onwarn(warning)
+}
 
 export default {
 	client: {
@@ -32,6 +48,7 @@ export default {
 				dedupe: ['svelte']
 			}),
 			commonjs(),
+			json(),
 
 			legacy && babel({
 				extensions: ['.js', '.mjs', '.html', '.svelte'],
@@ -74,7 +91,8 @@ export default {
 			resolve({
 				dedupe: ['svelte']
 			}),
-			commonjs()
+			commonjs(),
+			json(),
 		],
 		external: Object.keys(pkg.dependencies).concat(
 			require('module').builtinModules || Object.keys(process.binding('natives'))