+++ src/routes/_layout.svelte
@@ -1,3 +1,11 @@
+<script context="module">
+	import { isLoading, waitLocale } from 'svelte-i18n';
+
+  	export async function preload(page) {
+    	return waitLocale();
+  	}
+</script>
+
 <script>
 	import Header from '../components/Header.svelte';
 	import Footer from '../components/Footer.svelte';
@@ -8,6 +16,9 @@
     <meta name="description" content="Dividi la cuenta y todos a mano" />
 </svelte:head>
 
+{#if $isLoading}
+  	<div class="loading">Loading...</div>
+{:else}
 <div id="app">
 	<Header />
 
@@ -17,3 +28,4 @@
 
 	<Footer />
 </div>
+{/if}