+++ src/server.js
@@ -3,6 +3,8 @@ import polka from 'polka';
 import compression from 'compression';
 import * as sapper from '@sapper/server';
 
+import { i18nMiddleware } from './i18n.js';
+
 const { PORT, NODE_ENV } = process.env;
 const dev = NODE_ENV === 'development';
 
@@ -13,6 +15,7 @@ polka() // You can also use Express
         url,
 		compression({ threshold: 0 }),
 		sirv('static', { dev }),
+		i18nMiddleware(),
 		sapper.middleware()
 	)
 	.listen(PORT, err => {