+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 # La Cuenta Changelog
 
+## 4.0.0
+- [x] Added internationalization with [svelte-i18n](https://github.com/kaisermann/svelte-i18n)
+- [x] Added modal for "Delete All" confirmation
+
 ## 3.0.0
 - [x] New UI